[#162495476] update sign in endpoint for users
