--- src/usingJSObject/helpers/checks.js
@@ -1,15 +1,0 @@
-const checkLocation = (req, res, next) => {
-    const {
-      location,
-    } = req.body;
-    
-    if (location === '' || location === undefined) {
-      return res.status(400).json({
-        status: 400,
-        error: 'pls input location',
-      })
-    }
-    return next();
-  };
-  
-  export default checkLocation;
