--- src/usingJSObject/package.text
@@ -1,52 +1,0 @@
-{
-  "name": "iReporter",
-  "version": "1.0.0",
-  "description": "iReporter application provides a platform that can be used to report corruption or observation. Users can anonymously submit new reports, review existing reports, include location and track the status of problems of observations they have reported.",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon src/server.js --exec babel-node",
-    "build": "babel src -d dist",
-    "serve": "node dist/server.js",
-    "test": "nyc mocha --require babel-register src/test/redFlag-test.js --exit",
-    "coverage": "nyc --reporter=html --reporter=text --reporter=lcov mocha --reporter spec --require babel-register src/test/*.js --timeout 600000 --exit",
-    "coveralls": "npm run coverage -- --report lcovonly && cat coverage/lcov.info | coveralls",
-    "deploy": "gh-pages -d UI"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/femitj/iReporter.git"
-  },
-  "author": "Tijani Femi",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/femitj/iReporter/issues"
-  },
-  "homepage": "https://github.com/femitj/iReporter",
-  "engines": {
-    "node": "10.6.0",
-    "npm": "6.1.0"
-  },
-  "dependencies": {
-    "body-parser": "^1.18.3",
-    "dotenv": "^6.2.0",
-    "express": "^4.16.4",
-    "gh-pages": "^2.0.1",
-    "make-runnable": "^1.3.6",
-    "moment": "^2.22.2",
-    "pg": "^7.7.1",
-    "uuid": "^3.3.2"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-register": "^6.26.0",
-    "chai": "^4.2.0",
-    "chai-http": "^4.2.0",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.9.0",
-    "mocha": "^5.2.0",
-    "mocha-lcov-reporter": "^1.3.0",
-    "nodemon": "^1.18.7",
-    "nyc": "^13.1.0"
-  }
-}
